+++ crates/query/src/plan/type_join/type_join_many.rs
@@ -1020,7 +1020,7 @@ impl PlanNode for TypeJoinMany {
 
         obj.insert(
             "iterations".to_string(),
-            serde_json::json!(self.exec_info.iterations as u64),
+            serde_json::json!(self.exec_info.iterations),
         );
 
         let parent_execute = self.parent_plan.explain_execute();
